% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shapefiles.R
\name{load_shape_ecoregions}
\alias{load_shape_ecoregions}
\title{Loads a simplified shapefile for ICES Ecoregions.}
\usage{
load_shape_ecoregions()
}
\value{
sf file
}
\description{
Loads a simplified shapefile for ICES Ecoregions.
}
\note{
Can add some helpful information here
}
\examples{
\dontrun{
load_shape_ecoregions()
}



}
