% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shapefiles.R
\name{load_shape_europe}
\alias{load_shape_europe}
\title{Loads a simplified shapefile for EU coastline.}
\usage{
load_shape_europe()
}
\value{
sf file
}
\description{
Loads a simplified shapefile for EU coastline.
}
