% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive_map.R
\name{map_ecoregion}
\alias{map_ecoregion}
\title{Function to plot the intercative map of ecoregions. Taken from AdviceXplorer. https://github.com/ices-tools-dev/adviceXplorer/}
\usage{
map_ecoregion(shape_eco, eu_shape)
}
\arguments{
\item{shape_eco}{(ecoregions' shapefile)}

\item{eu_shape}{(europe's shapefile)}
}
\value{
leaflet object
}
\description{
Function to plot the intercative map of ecoregions. Taken from AdviceXplorer. https://github.com/ices-tools-dev/adviceXplorer/
}
\examples{
\dontrun{
map_ecoregion(shape_eco, eu_shape)
}
}
