% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive_map.R
\name{map_panel_server}
\alias{map_panel_server}
\title{Server side functionality for the landing page map panel}
\usage{
map_panel_server(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Server side functionality for the landing page map panel
}
