% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_commit_date.R
\name{extract_github_commit_date}
\alias{extract_github_commit_date}
\title{Extract the Last Commit Date of a Specific File from a GitHub Repository}
\usage{
extract_github_commit_date(region)
}
\value{
A character string in the format \code{"Last updated on: dd/mm/yyyy"} representing the last commit date,
        or \code{NULL} if the request fails or the structure is unexpected.
}
\description{
This function queries the GitHub API to retrieve the most recent commit date
If an error occurs, the function returns \code{NULL}.
}
\examples{
extract_github_commit_date()

}
