% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ggATAC}
\alias{ggATAC}
\title{## The 2 main functions: ATAC (modelling) and ggATAC (plotting)
Modelling is done with the [BRMS package](https://paul-buerkner.github.io/brms/)
Plotting is done using the [ggplot2 package](https://ggplot2.tidyverse.org)
data = dataset with two columns: year and value
year_start = first year that is be used when fitting the models
year_end_trend = last year that is used when fitting the overall trend model (using all observations). By default this is the last year of the series
year_end_OL = last year that is used when fitting the model that is used for forecasting. By default this will be the last year of data minus the n years of forecasts
n_forecasts = number of forecasts to the produced. By default 3.}
\usage{
ggATAC(results, backtransform = FALSE, trend = FALSE, width)
}
\description{
## The 2 main functions: ATAC (modelling) and ggATAC (plotting)
Modelling is done with the [BRMS package](https://paul-buerkner.github.io/brms/)
Plotting is done using the [ggplot2 package](https://ggplot2.tidyverse.org)
data = dataset with two columns: year and value
year_start = first year that is be used when fitting the models
year_end_trend = last year that is used when fitting the overall trend model (using all observations). By default this is the last year of the series
year_end_OL = last year that is used when fitting the model that is used for forecasting. By default this will be the last year of data minus the n years of forecasts
n_forecasts = number of forecasts to the produced. By default 3.
}
